Remplace maxBadgeToShow par disableBadgeDisplayedLimit qui enlève simplement la limite de badges affichés
