+++ apps/app/src/app/pages/collectivite/Indicateurs/detail/indicateur-sources.select.tsx
@@ -15,7 +15,7 @@ export const IndicateurSourcesSelect = ({ sourceFilter }: Props) => {
     <SelectMultiple
       isLoading={isLoading}
       dataTest="indicateurs.sources-select"
-      maxBadgesToShow={availableOptions.length}
+      disableBadgeDisplayedLimit
       options={availableOptions}
       values={filtresSource}
       onChange={({ values }) =>
