+++ apps/app/src/app/pages/collectivite/Referentiels/DownloadScore/download-score.modal.tsx
@@ -122,7 +122,7 @@ export const DownloadScoreModal = ({
                 values={selectedSnapshots ?? []}
                 onChange={setSelectedSnapshots}
                 options={snapshots}
-                maxBadgesToShow={2}
+                maxSelection={2}
                 aria-labelledby="versions-legend"
                 aria-describedby="selection-help"
                 aria-required="true"
