+++ packages/ui/src/design-system/Select/SelectMultiple.tsx
@@ -15,11 +15,6 @@ export type SelectMultipleProps = Omit<SelectProps, 'onChange' | 'values'> & {
   onChange: (args: SelectMultipleOnChangeArgs) => void;
   /** Valeurs sélectionnées */
   values?: OptionValue[];
-  /**
-   * Nombre maximum de badges à afficher
-   * @deprecated Peut entrainer des problèmes d'affichage pour des valeurs avec beaucoup de texte ou des selecteurs pas très large. Déconseillé de mettre une value supérieure à 1 tant que ce n'est pas géré correctement.
-   */
-  maxBadgesToShow?: number;
   optionsAreCaseSensitive?: boolean;
 };
 /**
@@ -30,7 +25,7 @@ export type SelectMultipleProps = Omit<SelectProps, 'onChange' | 'values'> & {
 export const SelectMultiple = ({
   values,
   onChange,
-  maxBadgesToShow,
+  disableBadgeDisplayedLimit,
   optionsAreCaseSensitive = true,
   ...props
 }: SelectMultipleProps) => {
@@ -60,7 +55,7 @@ export const SelectMultiple = ({
       }}
       multiple
       values={values}
-      maxBadgesToShow={maxBadgesToShow}
+      disableBadgeDisplayedLimit={disableBadgeDisplayedLimit}
       optionsAreCaseSensitive={optionsAreCaseSensitive}
     />
   );
